requirements.txt
fastapi==0.95.0
uvicorn==0.21.1

sqlalchemy==1.4.47
psycopg2-binary==2.9.5

alembic==1.10.2

python-dotenv==1.0.0

pytest==7.3.1
pytest-asyncio==0.21.0
python-multipart==0.0.6
